requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.4
cryptography==3.4.5
defusedxml==0.7.0rc2
Django==3.1.3
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
model-bakery==1.2.1
oauthlib==3.1.0
packaging==20.8
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.11.1
pytest-django==4.1.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.0
